requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
asyncio
pytz>=2023.3